refactor: memo editor components (#1625)
