+++ web/src/less/memo-editor.less
@@ -39,50 +39,6 @@
       > .action-btn {
         @apply flex flex-row justify-center items-center p-1 w-auto h-auto mr-1 select-none rounded cursor-pointer text-gray-600 dark:text-gray-200 hover:bg-gray-300 dark:hover:bg-zinc-800 hover:shadow;
 
-        &.tag-action {
-          @apply relative;
-
-          &:hover {
-            > .tag-list {
-              @apply flex;
-            }
-          }
-
-          > .tag-list {
-            @apply hidden flex-row justify-start items-start flex-wrap absolute top-6 left-0 mt-1 p-1 z-1 rounded w-52 h-auto max-h-48 overflow-y-auto font-mono shadow bg-zinc-200 dark:bg-zinc-600;
-
-            > .item-container {
-              @apply w-auto max-w-full truncate text-black dark:text-gray-300 cursor-pointer rounded text-sm leading-6 px-2 hover:bg-zinc-300 dark:hover:bg-zinc-700 shrink-0;
-            }
-
-            > .tip-text {
-              @apply w-auto text-sm text-gray-200 leading-6 px-2 cursor-default;
-            }
-          }
-        }
-
-        &.resource-btn {
-          @apply relative;
-
-          &:hover {
-            > .resource-action-list {
-              @apply flex;
-            }
-          }
-
-          > .resource-action-list {
-            @apply hidden flex-col justify-start items-start absolute top-6 left-0 mt-1 p-1 z-1 rounded w-auto overflow-auto font-mono shadow bg-zinc-200 dark:bg-zinc-600;
-
-            > .resource-action-item {
-              @apply w-full flex text-black dark:text-gray-300 cursor-pointer rounded text-sm leading-6 px-2 truncate hover:bg-zinc-300 dark:hover:bg-zinc-700 shrink-0;
-
-              > .icon-img {
-                @apply w-4 mr-1;
-              }
-            }
-          }
-        }
-
         > .icon-img {
           @apply w-5 h-5 mx-auto flex flex-row justify-center items-center;
         }
